+++ app/Models/User.php
@@ -6,6 +6,7 @@ namespace App\Models;
 
 use App\Contracts\Models\Viewable;
 use App\Enums\UserMailPreference;
+use App\Services\ParsableBio;
 use Database\Factories\UserFactory;
 use Filament\Models\Contracts\FilamentUser;
 use Filament\Panel;
@@ -20,6 +21,7 @@ use Illuminate\Notifications\DatabaseNotification;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\HtmlString;
 use Laravel\Fortify\TwoFactorAuthenticatable;
 
 /**
@@ -27,6 +29,7 @@ use Laravel\Fortify\TwoFactorAuthenticatable;
  * @property string|null $avatar
  * @property string $avatar_url
  * @property string|null $bio
+ * @property HtmlString $parsed_bio
  * @property Carbon $created_at
  * @property string $email
  * @property Carbon|null $email_verified_at
@@ -97,7 +100,7 @@ final class User extends Authenticatable implements FilamentUser, MustVerifyEmai
     /**
      * Get the user's bookmarks.
      *
-     * @return HasMany<Bookmark>
+     * @return HasMany<Bookmark, $this>
      */
     public function bookmarks(): HasMany
     {
@@ -107,7 +110,7 @@ final class User extends Authenticatable implements FilamentUser, MustVerifyEmai
     /**
      * Get the user's links.
      *
-     * @return HasMany<Link>
+     * @return HasMany<Link, $this>
      */
     public function links(): HasMany
     {
@@ -117,7 +120,7 @@ final class User extends Authenticatable implements FilamentUser, MustVerifyEmai
     /**
      * Get the user's questions sent.
      *
-     * @return HasMany<Question>
+     * @return HasMany<Question, $this>
      */
     public function questionsSent(): HasMany
     {
@@ -127,7 +130,7 @@ final class User extends Authenticatable implements FilamentUser, MustVerifyEmai
     /**
      * Get the user's questions received.
      *
-     * @return HasMany<Question>
+     * @return HasMany<Question, $this>
      */
     public function questionsReceived(): HasMany
     {
@@ -135,7 +138,7 @@ final class User extends Authenticatable implements FilamentUser, MustVerifyEmai
     }
 
     /**
-     * @return HasOne<Question>
+     * @return HasOne<Question, $this>
      */
     public function pinnedQuestion(): HasOne
     {
@@ -146,7 +149,7 @@ final class User extends Authenticatable implements FilamentUser, MustVerifyEmai
     /**
      * Get the user's followers.
      *
-     * @return BelongsToMany<User>
+     * @return BelongsToMany<User, $this>
      */
     public function followers(): BelongsToMany
     {
@@ -156,7 +159,7 @@ final class User extends Authenticatable implements FilamentUser, MustVerifyEmai
     /**
      * Get the user's following.
      *
-     * @return BelongsToMany<User>
+     * @return BelongsToMany<User, $this>
      */
     public function following(): BelongsToMany
     {
@@ -283,6 +286,14 @@ final class User extends Authenticatable implements FilamentUser, MustVerifyEmai
         }
 
         return $isCompanyVerified;
+    }
+
+    /**
+     * Get the user's bio attribute.
+     */
+    public function getParsedBioAttribute(): HtmlString
+    {
+        return new HtmlString((new ParsableBio)->parse((string) $this->bio));
     }
 
     /**