+++ app/Queries/Feeds/QuestionsFollowingFeed.php
@@ -7,6 +7,7 @@ namespace App\Queries\Feeds;
 use App\Models\Question;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Query\Builder as QueryBuilder;
 use Illuminate\Support\Facades\DB;
 
 final readonly class QuestionsFollowingFeed
@@ -27,7 +28,12 @@ final readonly class QuestionsFollowingFeed
     {
         $followQueryClosure = function (Builder $query): void {
             $query->where('to_id', $this->user->id)
-                ->orWhereIn('to_id', $this->user->following()->select('users.id'));
+                ->orWhereExists(function (Builder|QueryBuilder $query): void {
+                    $query->select(DB::raw(1))
+                        ->from('followers')
+                        ->whereColumn('user_id', 'to_id')
+                        ->where('follower_id', $this->user->id);
+                });
         };
 
         return Question::query()
@@ -36,15 +42,12 @@ final readonly class QuestionsFollowingFeed
                 'root as showRoot' => $followQueryClosure,
                 'parent as showParent' => $followQueryClosure,
             ])
-            ->withAggregate('to as username', 'username')
-            ->withAggregate('parent as grand_parent_id', 'parent_id')
+            ->with('root:id,to_id', 'root.to:id,username', 'parent:id,parent_id')
             ->whereNotNull('answer')
             ->where('is_reported', false)
             ->where('is_ignored', false)
             ->where($followQueryClosure)
-            ->where(function (Builder $query): void {
-                $query->whereNull('parent_id')->orWhere('showParent', true)->orWhere('showRoot', true);
-            })
+            ->havingRaw('parent_id IS NULL or showRoot = 1 or showParent = 1')
             ->groupBy(DB::Raw('IFNULL(root_id, id)'))
             ->orderByDesc(DB::raw('MAX(`updated_at`)'));
     }