+++ app/Queries/Feeds/RecentQuestionsFeed.php
@@ -39,12 +39,7 @@ final readonly class RecentQuestionsFeed
                 })->orderByDesc('updated_at');
             }, function (Builder $query): void {
                 $query->addSelect('root_id', 'parent_id')
-                    ->withAggregate('to as username', 'username')
-                    ->withAggregate('parent as grand_parent_id', 'parent_id')
-                    ->where(function (Builder $query): void {
-                        $query->whereNull('root_id')
-                            ->orHas('root');
-                    })
+                    ->with('root.to:username,id', 'root:id,to_id', 'parent:id,parent_id')
                     ->groupBy(DB::Raw('IFNULL(root_id, id)'))
                     ->orderByDesc(DB::raw('MAX(`updated_at`)'));
             });