+++ resources/views/layouts/navigation.blade.php
@@ -1,5 +1,5 @@
 @php
-    $navClasses = 'fixed z-50 inset-0 h-16 flex md:justify-end md:px-4 ';
+    $navClasses = 'fixed z-50 inset-0 md:inset-auto md:right-0 h-16 flex md:justify-end md:px-4 ';
     $navClasses .= auth()->check() ? ' justify-center' : ' justify-end px-4';
 @endphp