+++ resources/views/livewire/home/users.blade.php
@@ -27,6 +27,7 @@
                         data-parent=true
                         x-data="clickHandler"
                         x-on:click="handleNavigation($event)"
+                        wire:key="user-{{ $user->id }}"
                     >
                         <div class="group flex items-center gap-3 rounded-2xl border dark:border-slate-900 border-slate-200 dark:bg-slate-950 bg-slate-50 dark:bg-opacity-80 p-4 transition-colors dark:hover:bg-slate-900 hover:bg-slate-100">
                             <figure class="{{ $user->is_company_verified ? 'rounded-md' : 'rounded-full' }} h-12 w-12 flex-shrink-0 overflow-hidden bg-slate-800 transition-opacity group-hover:opacity-90">
@@ -68,6 +69,7 @@
                                 :isFollower="auth()->check() && $user->is_follower"
                                 :isFollowing="auth()->check() && $user->is_following"
                                 class="ml-auto"
+                                wire:key="follow-button-{{ $user->id }}"
                             />
                         </div>
                     </li>