+++ tests/Unit/Services/ContentProvidersTest.php
@@ -25,6 +25,10 @@ test('links', function (string $content, string $parsed) {
     expect($provider->parse($content))->toBe($parsed);
 })->with([
     [
+        'content' => 'http://example.com/',
+        'parsed' => '<a data-navigate-ignore="true" class="text-blue-500 hover:underline hover:text-blue-700 cursor-pointer" target="_blank" href="http://example.com/">example.com</a>',
+    ],
+    [
         'content' => 'https://example.com/',
         'parsed' => '<a data-navigate-ignore="true" class="text-blue-500 hover:underline hover:text-blue-700 cursor-pointer" target="_blank" href="https://example.com/">example.com</a>',
     ],
@@ -190,6 +194,25 @@ test('links within <pre> and <code> blocks are ignored', function (string $conte
     ],
 ]);
 
+test('only http or https urls are converted to links', function (string $content, string $parsed) {
+    $provider = new App\Services\ParsableContentProviders\LinkProviderParsable();
+
+    expect($provider->parse($content))->toBe($parsed);
+})->with([
+    [
+        'content' => 'example.com will not be a link but https://example.com will',
+        'parsed' => 'example.com will not be a link but <a data-navigate-ignore="true" class="text-blue-500 hover:underline hover:text-blue-700 cursor-pointer" target="_blank" href="https://example.com">example.com</a> will',
+    ],
+    [
+        'content' => 'www.example.com',
+        'parsed' => 'www.example.com',
+    ],
+    [
+        'content' => 'blade.php',
+        'parsed' => 'blade.php',
+    ],
+]);
+
 test('mentions within <pre> and <code> blocks are ignored', function (string $content, string $parsed) {
     $provider = new App\Services\ParsableContentProviders\MentionProviderParsable();
 
@@ -243,6 +266,19 @@ test('code', function (string $content) {
             echo "Hello, World!";
             ```
             EOL,
+    ],
+    /*
+        The below example tests that the code block is still parsed correctly even if there
+        is a space after the language. Sonarlint flags up a 'useless space' error
+        so we need to use a str_replace to add a space after the language programatically.
+    */
+    [
+        'content' => str_replace('```php', '```php ', <<<'EOL'
+            ```php
+            echo "Hello, World!";
+            ```
+            EOL
+        ),
     ],
 ]);